TCIO-146 Developed review controller layer logic
TCIO-135 Developed reviews processing

Also updated Reviews Service and Security logic in controllers and overrided equals and hashCode in Figure, User, Review classes

+++ src/main/java/com/teamChallenge/controller/BannerController.java
@@ -3,8 +3,11 @@ package com.teamChallenge.controller;
 import com.teamChallenge.dto.request.BannerRequestDto;
 import com.teamChallenge.dto.response.BannerResponseDto;
 import com.teamChallenge.entity.banner.BannerService;
+import com.teamChallenge.entity.user.UserServiceImpl;
 import com.teamChallenge.exception.CustomErrorResponse;
 import com.teamChallenge.exception.LogEnum;
+import com.teamChallenge.exception.exceptions.generalExceptions.UnauthorizedAccessException;
+import com.teamChallenge.security.AuthUtil;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -17,6 +20,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
 import java.util.List;
 
 @RestController
@@ -28,6 +32,9 @@ public class BannerController {
     public static final String URI_BANNER_WITH_ID = "/{id}";
 
     private final BannerService bannerService;
+    private final AuthUtil authUtil;
+
+    private final String secReq = "BearerAuth";
 
     @GetMapping
     @Operation(summary = "Get all banners")
@@ -62,7 +69,6 @@ public class BannerController {
     }
 
     @PostMapping
-    @SecurityRequirement(name = "BearerAuth")
     @Operation(description = "create a banner")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "201", description = "Created the banner",
@@ -74,14 +80,16 @@ public class BannerController {
                             schema = @Schema(implementation = CustomErrorResponse.class))}
             ),
     })
-    public BannerResponseDto create(@RequestBody BannerRequestDto bannerRequestDto) {
+    @SecurityRequirement(name = secReq)
+    public BannerResponseDto create(@RequestBody BannerRequestDto bannerRequestDto, Principal principal) throws UnauthorizedAccessException {
+        authUtil.validateAdminRole(principal);
+
         BannerResponseDto bannerResponseDto = bannerService.create(bannerRequestDto);
         log.info("{}: Banner (id: {}) has been added", LogEnum.CONTROLLER, bannerResponseDto.id());
         return bannerResponseDto;
     }
 
     @PutMapping(URI_BANNER_WITH_ID)
-    @SecurityRequirement(name = "BearerAuth")
     @Operation(description = "update a banner by id")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "Updated the banner",
@@ -97,7 +105,10 @@ public class BannerController {
                             schema = @Schema(implementation = CustomErrorResponse.class))
                     })
     })
-    public BannerResponseDto update(@PathVariable String id, @RequestBody BannerRequestDto bannerRequestDto) {
+    @SecurityRequirement(name = secReq)
+    public BannerResponseDto update(@PathVariable String id, @RequestBody BannerRequestDto bannerRequestDto, Principal principal) throws UnauthorizedAccessException {
+        authUtil.validateAdminRole(principal);
+
         BannerResponseDto bannerResponseDto = bannerService.update(id, bannerRequestDto);
         log.info("{}: Banner (id: {}) has been updated", LogEnum.CONTROLLER, id);
         return bannerResponseDto;
@@ -116,7 +127,9 @@ public class BannerController {
                             schema = @Schema(implementation = CustomErrorResponse.class))
                     })
     })
-    public void delete(@PathVariable String id) {
+    public void delete(@PathVariable String id, Principal principal) throws UnauthorizedAccessException {
+        authUtil.validateAdminRole(principal);
+
         bannerService.delete(id);
         log.info("{}: Banner (id: {}) has been deleted", LogEnum.CONTROLLER, id);
     }
